fixed changes requested
