+++ conf/conf.toml
@@ -1,7 +1,7 @@
 [nope]
 long_enter = -60
 long_exit  = -30
-short_enter = 0
+short_enter = 30
 short_exit  = 15
 
 [debug]
