+++ utils/util.py
@@ -1,0 +1,10 @@
+
+# From thetagang
+# https://github.com/brndnmtthws/thetagang
+def while_n_times(pred, body, remaining):
+    if remaining <= 0:
+        raise RuntimeError(
+            "Exhausted retries waiting on predicate. This shouldn't happen.")
+    if pred() and remaining > 0:
+        body()
+        while_n_times(pred, body, remaining - 1)
